Rename agents.dhcp to agents.impl.dhcp. (Create implementation package level.)
